take table fetching into the table-data module
